chore: Refactor src/i18n folder
