+++ template/src/plugins/i18n.js
@@ -6,8 +6,8 @@ export default ({ app, Vue }) => {
 
   // Set i18n instance on app
   app.i18n = new VueI18n({
-    locale: 'en',
-    fallbackLocale: 'en',
+    locale: 'en-us',
+    fallbackLocale: 'en-us',
     messages
   })
 }
